Clean up components structure a bit
