+++ components/CodeExample/index.tsx
@@ -2,21 +2,22 @@ import TopBar from './TopBar';
 import React, { useEffect, useState } from 'react'
 import { CopyBlock, dracula } from 'react-code-blocks';
 import codeExamplesFactory from './codeExamplesFactory';
+import { useRouter } from 'next/router';
 
 // TODO: The same like with Markdown here we may want to rename this
 // TODO: This is dummy sync function for testing purposes we have to change it the way example will be rendered with mounting of component
 
-type Props = {
-    codeStructure: string | string[] | undefined;
-}
 
 type CodeExample = {
     [key: string]: string;
 }
 
-function CodeExample({ codeStructure }: Props) {
+function CodeExample() {
+    const router = useRouter();
+    // TODO: consider changing name 'dataStructure' because in the future it might be also an algorithm 
+    const { dataStructure } = router.query;
 
-    const codeExamples: CodeExample | string | undefined = codeExamplesFactory(codeStructure);
+    const codeExamples: CodeExample | string | undefined = codeExamplesFactory(dataStructure);
 
     useEffect(() => {
         if (codeExamples) {
